refactor CSV importer to use new module structure and refactor test imports
