--- backend/apps/imports/csv_importer/__init__.py
@@ -1,5 +1,0 @@
-"""Legacy CSV importer package for Viernulvier archive data."""
-
-from .legacy_csv import detect_legacy_csv_kind, import_bundled_legacy_csv_files, import_legacy_csv_file
-
-__all__ = ["detect_legacy_csv_kind", "import_bundled_legacy_csv_files", "import_legacy_csv_file"]
