+++ backend/tests/imports_csv/test_legacy_csv_productions.py
@@ -4,7 +4,7 @@ import pytest
 
 from apps.genres.models import Genre, GenreTranslation, GenreUseAs
 from apps.import_log.models import ImportLog
-from apps.imports.csv_importer import import_legacy_csv_file
+from apps.imports.csv_importer.legacy_csv_sync import import_legacy_csv_file
 from apps.languages.models import Language
 from apps.locations.models import Hall
 from apps.productions.models import Production, ProductionGenre
